IHARP/submission/requirements.txt
torch==2.3.1
torchvision==0.18.1
matplotli==3.10.0
pillow==10.2.0
numpy==2.2.0